+++ packages/cli/plugin-data-loader/src/server/index.ts
@@ -3,10 +3,7 @@ import path from 'path';
 import fs from 'fs';
 import type { ServerPlugin } from '@modern-js/server-core';
 import type { ServerRoute } from '@modern-js/types';
-import {
-  MAIN_ENTRY_NAME,
-  SERVER_BUNDLE_DIRECTORY,
-} from '@modern-js/utils/constants';
+import { MAIN_ENTRY_NAME, SERVER_BUNDLE_DIRECTORY } from '@modern-js/utils';
 import { matchEntry, ServerContext } from '../common/utils';
 
 export default (): ServerPlugin => ({