+++ tests/integration/routes/tests/index.test.ts
@@ -1,7 +1,6 @@
 /* eslint-disable max-lines */
 import path from 'path';
-import { fs } from '@modern-js/utils';
-import { ROUTE_MANIFEST_FILE } from '@modern-js/utils/constants';
+import { fs, ROUTE_MANIFEST_FILE } from '@modern-js/utils';
 import { ROUTE_MANIFEST } from '@modern-js/utils/universal/constants';
 import type {
   // Browser,