+++ packages/builder/plugin-swc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/builder-plugin-swc
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/builder-shared@2.18.0
+- @modern-js/builder-plugin-swc-base@2.18.0
+- @modern-js/utils@2.18.0
+
 ## 2.17.1
 
 ### Patch Changes