+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-changeset
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.18.0
+- @modern-js/utils@2.18.0
+
 ## 2.17.1
 
 ### Patch Changes