+++ packages/generator/generators/router-v5-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/router-v5-generator
 
+## 3.1.13
+
 ## 3.1.12
 
 ## 3.1.11