+++ packages/review/eslint-config-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js-app/eslint-config
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/babel-preset-app@2.18.0
+
 ## 2.17.1
 
 ### Patch Changes