+++ packages/server/server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/server
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/prod-server@2.18.0
+- @modern-js/server-utils@2.18.0
+- @modern-js/types@2.18.0
+- @modern-js/utils@2.18.0
+
 ## 2.17.1
 
 ### Patch Changes